feat: Refactor/Restructure API views
Created 'views' module with app named files for storing their respective
views
