+++ api/views/employee.py
@@ -1,13 +1,19 @@
 from rest_framework import viewsets
 
-from employee.models import Employee, Department, Designation, Location
-from hardware.models import Laptop, LaptopBrand, Building
-from finance.models import Payment
+from employee.models import (
+    Employee,
+    Department,
+    Designation,
+    Location
+)
+from api.serializers import (
+    EmployeeSerializer,
+    DepartmentSerializer,
+    DesignationSerializer,
+    LocationSerializer,
+)
+from api.custom_pagination import FullResultsSetPagination
 
-from .serializers import *
-from .custom_pagination import FullResultsSetPagination
-
-# Employee Viewsets
 class EmployeeViewSet(viewsets.ModelViewSet):
     queryset = Employee.objects.all()
     serializer_class = EmployeeSerializer
@@ -33,21 +39,3 @@ class DesignationViewSet(viewsets.ModelViewSet):
 class LocationViewSet(viewsets.ModelViewSet):
     queryset = Location.objects.all()
     serializer_class = LocationSerializer
-
-# Hardware Viewsets
-class LaptopViewSet(viewsets.ModelViewSet):
-    queryset = Laptop.objects.all()
-    serializer_class = LaptopSerializer
-
-class LaptopBrandViewSet(viewsets.ModelViewSet):
-    queryset = LaptopBrand.objects.all()
-    serializer_class = LaptopBrandSerializer
-
-class BuildingViewSet(viewsets.ModelViewSet):
-    queryset = Building.objects.all()
-    serializer_class = BuildingSerializer
-
-# Finance Viewsets
-class PaymentViewSet(viewsets.ModelViewSet):
-    queryset = Payment.objects.all()
-    serializer_class = PaymentSerializer
