+++ api/views/finance.py
@@ -1,0 +1,8 @@
+from rest_framework import viewsets
+
+from finance.models import Payment
+from api.serializers import PaymentSerializer
+
+class PaymentViewSet(viewsets.ModelViewSet):
+    queryset = Payment.objects.all()
+    serializer_class = PaymentSerializer
